Update index.html
Changed the padding to certain section to adjust the navbar direction to that clicked section. Since when clicking on reviews section it would bring user midway rather at the start of the section. So I adjusted the section appropriately for all section. Most likely a more effective method or better consistency.